+++ packages/botonic-api/src/dispatchers/index.ts
@@ -1,0 +1,11 @@
+import { Environments } from '..'
+import { AWSDispatchers } from './aws-dispatchers'
+import { LocalDispatchers } from './local-dispatchers'
+
+export function dispatchersFactory(env, handlers) {
+  // @ts-ignore
+  if (ENV === Environments.LOCAL) {
+    return new LocalDispatchers(handlers)
+  }
+  return new AWSDispatchers()
+}
