+++ packages/botonic-api/src/dispatchers/local-dispatchers.ts
@@ -1,0 +1,10 @@
+export class LocalDispatchers {
+  localDispatchers
+
+  constructor(dispatchers) {
+    this.localDispatchers = dispatchers
+  }
+  async run(handlerName, params) {
+    await this.localDispatchers[handlerName](params)
+  }
+}
