+++ packages/create-botonic-app/dev-template/api/src/handlers/botExecutor.js
@@ -1,11 +1,15 @@
 // eslint-disable-next-line filenames/match-regex
 import { botExecutorHandlerFactory } from '@botonic/api/src/handlers/bot-executor'
 import { dataProviderFactory } from '@botonic/core/lib/esm/data-provider'
+import { dispatchersFactory } from '@botonic/api/src/dispatchers'
+
 import { app as bot } from 'bot'
 
-import { dispatchers } from '.'
+import sender from './sender'
 
 const dataProvider = dataProviderFactory(process.env.DATA_PROVIDER_URL)
+
+const dispatchers = dispatchersFactory(ENV, { sender })
 
 async function botExecutor({ userId, input, session, botState, websocketId }) {
   const output = await bot.input({
