--- packages/create-botonic-app/dev-template/api/src/handlers/index.js
@@ -1,11 +1,0 @@
-import { Dispatchers } from '@botonic/api/src/handlers'
-import { Environments } from '@botonic/api/src/index'
-
-export const dispatchers =
-  // eslint-disable-next-line no-undef
-  ENV === Environments.LOCAL
-    ? new Dispatchers({
-        botExecutor: require('./botExecutor').default,
-        sender: require('./sender').default,
-      })
-    : new Dispatchers(undefined)
