+++ packages/create-botonic-app/dev-template/api/src/rest/index.js
@@ -1,7 +1,12 @@
 import { restServerFactory } from '@botonic/api/src/rest'
+import { dispatchersFactory } from '@botonic/api/src/dispatchers/index'
 
-import { dispatchers } from '../handlers'
+import botExecutor from '../handlers/botExecutor'
+import sender from '../handlers/sender'
+
 import { getApp } from './app'
+
+export const dispatchers = dispatchersFactory(ENV, { botExecutor, sender })
 
 export default restServerFactory({
   // eslint-disable-next-line no-undef
